requirements.txt
mysql-connector-python==8.0.33
pandas==2.0.3
numpy==1.24.3
python-dateutil==2.8.2